waheed_ahmad/sprint2/requirements.txt

aws-cdk.core==1.135.0
aws-cdk.aws_lambda==1.135.0
aws-cdk.aws_events==1.135.0
aws-cdk.aws_iam==1.135.0
aws-cdk.aws_events_targets==1.135.0
aws-cdk.aws_codepipeline==1.135.0
aws-cdk.aws_codepipeline_actions==1.135.0
aws-cdk.pipelines==1.135.0
pytest
boto3